Refactor purchase return validation into shared concern
